% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRICED_bboot.R
\name{PRICED_bboot}
\alias{PRICED_bboot}
\title{Perform inference for the PRICED model using the Bayesian bootstrap for inference}
\usage{
PRICED_bboot(
  formula_prevalence,
  formula_incidence,
  formula_clearance,
  data,
  prior_prevalence = list(location = 0, scale = 2.5, autoscale = TRUE),
  prior_incidence = list(location = 0, scale = 2.5, autoscale = TRUE),
  prior_clearance = list(location = 0, scale = 2.5, autoscale = TRUE),
  sensitivity = 0.75,
  specificity = 0.98,
  n_bootstraps = 250,
  cl,
  verbose = TRUE
)
}
\arguments{
\item{formula_prevalence}{Right handed formula of the form  ~ x_1 + x_2 + ...}

\item{formula_incidence}{Right handed formula of the form  ~ x_1 + x_2 + ...}

\item{formula_clearance}{Formula of the form y ~ x_1 + x_2 + ... + (time_variable | subject_ID)}

\item{data}{Data frame or tibble}

\item{prior_clearance, prior_incidence, prior_beta_pr}{a list giving the normal hyperparameters location and scale (i.e., sd).
Additionally, autoscale can be set to TRUE, in which case the scale will be
divided by the standard deviation of the covariates for each regression coefficient.}

\item{sensitivity}{Diagnostic test sensitivity}

\item{specificity}{Diagnostic test specificity}

\item{verbose}{logical.  Whether to print messages as to where the model fitting algorithm is.}

\item{n_draws}{integer, number of posterior draws to obtain (not used if method = "normal")}

\item{seed}{a single value, interpreted as an integer.  Used for replication purposes.}
}
\value{
Object of class 'PRICED_boot' which has the following elements:
\itemize{
\item t0 The regression coefficients using the original sample
\item t A matrix where each row is a posterior sample and each column is a regression
coefficient
\item R  The number of bootstrap samples
\item formula_prevalence, formula_incidence, formula_clearance
}
}
\description{
Obtain posterior samples based on data fit to
the PRICED model, a set of regression models for
CLearance, Incidence, and Prevalence, accounting for
diagnostic errors.  Bayesian bootstrapping is used to get
posterior samples of the regression coefficients.
}
\examples{
\dontrun{
PRICED_data =
  simulate_PRICED(seed = 1,N_subj = 500, N_time = 10)
cl = makeCluster(detectCores())
PRICED_fit_parallel = PRICED_bboot(formula_prevalence = ~ x1 + x2 + x3,
                                   formula_incidence = ~ x1 + x2 + x3,
                                   formula_clearance = p_observed ~ x1 + x2 + x3 + (time | subject),
                                   data = PRICED_data$data[[1]],
                                   prior_prevalence = list(location = 0, scale = 2.5, autoscale = TRUE),
                                   prior_incidence = list(location = 0, scale = 2.5, autoscale = TRUE),
                                   prior_clearance = list(location = 0, scale = 2.5, autoscale = TRUE),
                                   sensitivity = 0.75,
                                   specificity = 0.99,
                                   n_bootstraps = detectCores() * 20,
                                   cl,
                                   verbose = TRUE)
stopCluster(cl)
summary(PRICED_fit_parallel)
}


}
