% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_PRICED.R
\name{simulate_PRICED}
\alias{simulate_PRICED}
\title{Simulate data from the PRICED model}
\usage{
simulate_PRICED(
  nsim = 1,
  seed = NULL,
  N_subj = 250,
  N_time = 5,
  bl_prevalence = 0.05,
  bl_clearance_rate = 1/3,
  sensitivity = 0.75,
  specificity = 0.98,
  beta_P_slopes = c(2, -2, 0),
  beta_I_slopes = c(2, -2, 0),
  beta_C_slopes = c(-2, 2, 0)
)
}
\arguments{
\item{nsim}{integer. The number of simulated data sets to create.}

\item{seed}{a single value, interpreted as an integer.  Used for replication purposes.}

\item{N_subj}{positive integer. Number of subjects in data sets}

\item{N_time}{Either positive integer or vector of integers of length N_subj.
Number of time points per subject.}

\item{bl_prevalence}{positive numeric, giving the
baseline prevalences for each pathogen type.}

\item{bl_clearance_rate}{positive numeric, giving the overall rate of clearance}

\item{sensitivity}{numeric, giving diagnostic test sensitivity}

\item{specificity}{numeric, giving diagnostic test specificity}

\item{beta_P_slopes, }{beta_I_slopes, beta_C_slopes numeric of length 3 corresponding
to the binary covariate and two continuous covariates respectively.}
}
\description{
Simulate data from the PRICED model
}
