% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PRICED_boot.R
\name{summary.PRICED_boot}
\alias{summary.PRICED_boot}
\title{Summarizing PRICED fits}
\usage{
\method{summary}{PRICED_boot}(
  object,
  CI_level = 0.95,
  estimator = c("mean", "median", "mode"),
  CI_method = c("HPD", "central")[1],
  print = TRUE,
  burnin = 0
)
}
\arguments{
\item{object}{an object of class "PRICED_boot"}

\item{CI_level}{level for credible intervals}

\item{estimator}{character, either "mean", "median", or "mode" for respective posterior point estimate.}

\item{CI_method}{character, either "HPD" (highest posterior density) or "central".}

\item{print}{logical}
}
\description{
summary method for class "PRICED"
}
\examples{
\dontrun{
 PRICED_fit_parallel = PRICED_bboot(formula_prevalence = ~ x1 + x2 + x3,
                                    formula_incidence = ~ x1 + x2 + x3,
                                    formula_clearance = p_observed ~ x1 + x2 + x3 + (time | subject),
                                    data = PRICED_data$data[[1]],
                                    prior_prevalence = list(location = 0, scale = 2.5, autoscale = TRUE),
                                    prior_incidence = list(location = 0, scale = 2.5, autoscale = TRUE),
                                   prior_clearance = list(location = 0, scale = 2.5, autoscale = TRUE),
                                   sensitivity = 0.75,
                                   specificity = 0.99,
                                   n_bootstraps = detectCores() * 20,
                                   cl,
                                   verbose = TRUE)
summary(PRICED_fit_parallel)
}

}
