% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PRICED.R
\name{summary.PRICED}
\alias{summary.PRICED}
\title{Summarizing PRICED fits}
\usage{
\method{summary}{PRICED}(
  object,
  CI_level = 0.95,
  estimator = c("mean", "median"),
  CI_method = c("HPD", "central")[1],
  print = TRUE,
  burnin = 0
)
}
\arguments{
\item{object}{an object of class "PRICED"}

\item{CI_level}{level for credible intervals}

\item{estimator}{character, either "mean" or "median" for respective posterior point estimate. Ignored if method = "normal"}

\item{CI_method}{character, either "HPD" (highest posterior density) or "central".  Ignored if method = "normal"}

\item{print}{logical}
}
\description{
summary method for class "PRICED"
}
\examples{
PRICED_data =
simulate_PRICED(seed = 2023,N_subj = 500, N_time = 10)

PRICED_fit1 =
  PRICED(formula_prevalence = ~ x1,
         formula_incidence = ~ x1 + x3,
         formula_clearance = p_observed ~ x1 + x2 + (time | subject),
         PRICED_data$data[[1]],
         prior_prevalence = list(location = 0, scale = 2.5, autoscale = TRUE),
         prior_incidence = list(location = 0, scale = 2.5, autoscale = TRUE),
         prior_clearance = list(location = 0, scale = 2.5, autoscale = TRUE),
         method = "normal",
         seed = 1,
         sensitivity = 0.75,
         specificity = 0.98,
         verbose = TRUE)
summary(PRICED_fit1)

set.seed(123)
PRICED_fit2 =
  PRICED(formula_prevalence = ~ x1,
         formula_incidence = ~ x1 + x3,
         formula_clearance = p_observed ~ x1 + x2 + (time | subject),
         PRICED_data$data[[1]],
         prior_prevalence = list(location = 0, scale = 2.5, autoscale = TRUE),
         prior_incidence = list(location = 0, scale = 2.5, autoscale = TRUE),
         prior_clearance = list(location = 0, scale = 2.5, autoscale = TRUE),
         method = "adaptMCMC",
         n_draws = 5e3,
         seed = 1,
         sensitivity = 0.75,
         specificity = 0.98,
         verbose = TRUE,
         acc.rate = 0.234)
summary(PRICED_fit2)
}
